# api/openapi/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/spatial-feature": {
      "get": {
        "operationId": "SpatialFeatureController_getForProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpatialFeaturePublicResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "spatial-feature"
        ]
      }
    },
    "/api/spatial-feature/bcgw-extract": {
      "get": {
        "operationId": "SpatialFeatureController_getBcgwExtract",
        "parameters": [
          {
            "name": "version",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpatialFeatureBcgwResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "spatial-feature"
        ]
      }
    },
...
  },
  "info": {
    "title": "FOM API",
    "description": "API for FOM backend",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SubmissionTypeCode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "description"
        ]
      },
      "FeatureTypeCode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "description"
        ]
      },
      "SpatialFeaturePublicResponse": {
        "type": "object",
        "properties": {
          "featureType": {
            "$ref": "#/components/schemas/FeatureTypeCode"
          },
          "featureId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "geometry": {
            "type": "object",
            "description": "Format: GeoJSON Geometry object"
          },
          "centroid": {
            "type": "object",
            "description": "Format: GeoJSON Point"
          },
          "plannedDevelopmentDate": {
            "type": "string",
            "description": "Format: YYYY-MM-DD"
          },
          "plannedAreaHa": {
            "type": "number"
          },
          "plannedLengthKm": {
            "type": "number"
          },
          "submissionType": {
            "$ref": "#/components/schemas/SubmissionTypeCode"
          }
        },
        "required": [
          "featureType",
          "featureId",
          "name",
          "geometry",
          "centroid",
          "plannedDevelopmentDate",
          "plannedAreaHa",
          "plannedLengthKm",
          "submissionType"
        ]
      },
      "SpatialFeatureBcgwResponse": {
        "type": "object",
        "properties": {
          "fomId": {
            "type": "number"
          },
          "featureType": {
            "type": "string",
            "enum": [
              "cut_block",
              "road_section",
              "retention_area"
            ]
          },
          "featureId": {
            "type": "number"
          },
          "name": {
            "type": "string",
            "description": "Feature object name",
            "example": "My cut block 2311"
          },
          "fspHolderName": {
            "type": "string",
            "description": "Maximum length is expected to be \u003C= 100"
          },
          "lifecycleStatus": {
            "type": "string",
            "enum": [
              "Proposed",
              "Final"
            ]
          },
          "createDate": {
            "type": "string",
            "description": "Format: YYYY-MM-DD",
            "example": "2021-03-31"
          },
          "geometry": {
            "type": "object",
            "description": "Format: GeoJSON Geometry as per RFC 7946 https://datatracker.ietf.org/doc/html/rfc7946#section-3.1.",
            "example": "Cut Block / Retention Area Example: {\"type\":\"Polygon\",\"coordinates\":[[[-121.652853481,51.67951884],[-121.499131212,51.763467947],[-121.363845891,51.668139701],[-121.373301212,51.578595415],[-121.661722435,51.589952346],[-121.652853481,51.67951884]]]} Road Section Example: {\"type\":\"LineString\",\"coordinates\":[[-124.667049923,52.651503193],[-124.739768473,52.697146151],[-124.816364174,52.60827305]]}"
          },
          "plannedDevelopmentDate": {
            "type": "string",
            "description": "Format: YYYY-MM-DD",
            "example": "2021-03-31"
          },
          "plannedAreaHa": {
            "type": "number",
            "example": "1.2345"
          },
          "plannedLengthKm": {
            "type": "number",
            "example": "1.2345"
          }
        },
        "required": [
          "fomId",
          "featureType",
          "featureId",
          "name",
          "fspHolderName",
          "lifecycleStatus",
          "createDate",
          "geometry",
          "plannedDevelopmentDate",
          "plannedAreaHa",
          "plannedLengthKm"
        ]
      },
...
    }
  }
}